fixup! consolidate MenuVariants tests into Menu.test.tsx
